+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.9.0 (2024-05-09)
 - [BREAKING] removed `group_vector_elements()` utility function (#282).
+- [BREAKING] removed `FieldElement::zeroed_vector()` function (#282).
 
 ## 0.8.4 (2024-03-28) - `math` crate only
 * Added more to/from conversions for `f64` field (#268).